Simplify defer type; separate tests
