+++ packages/inngest/src/test/integration/middleware/useCases/encoding.test.ts
@@ -104,6 +104,7 @@ test("base64 encoding/decoding middleware", async () => {
     ctx: {
       attempt: 0,
       maxAttempts: 1,
+      defer: expect.any(Function),
       event: expect.any(Object),
       events: [expect.any(Object)],
       group: expect.any(Object),
