+++ packages/inngest/src/test/integration/utils.ts
@@ -166,6 +166,7 @@ const anyEvent = {
 
 export const anyContext = {
   attempt: expect.any(Number),
+  defer: expect.any(Function),
   event: anyEvent,
   events: [anyEvent],
   group: expect.any(Object),
@@ -241,4 +242,13 @@ export function matrixCheckpointing(
       test(`checkpointing: ${checkpointing}`, () => fn(checkpointing));
     }
   });
+}
+
+export function spyLogger() {
+  return {
+    info: vi.fn(),
+    warn: vi.fn(),
+    error: vi.fn(),
+    debug: vi.fn(),
+  };
 }
